# Conflicts:
#   runnershi_fe/RunnersHi/app/src/main/res/layout/fragment_room.xml

+++ runnershi_fe/RunnersHi/app/src/main/java/com/A306/runnershi/Model/Run.kt
@@ -17,7 +17,7 @@ data class Run(
     // 평균 속력
     var avgSpeed: Float = 0f,
     // 거리
-    var distance: Int = 0,
+    var distance: Float = 0f,
     // 시간
     var time: String? = null,
     // 페이스
